% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{all_elements_same_cpp_wrap}
\alias{all_elements_same_cpp_wrap}
\title{Tests if a vector has all elements the same}
\usage{
all_elements_same_cpp_wrap(w)
}
\arguments{
\item{w}{The vector to be queried}
}
\value{
A boolean if it has all same elements
}
\description{
Tests if a vector has all elements the same
}
\author{
Adam Kapelner
}
