% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_randomization.R
\name{complete_randomization_with_balanced_ns}
\alias{complete_randomization_with_balanced_ns}
\title{Implements complete balanced randomization}
\usage{
complete_randomization_with_balanced_ns(n, r)
}
\arguments{
\item{n}{number of observations}

\item{r}{number of randomized designs you would like}
}
\value{
a matrix where each column is one of the \code{r} designs
}
\description{
Implements complete balanced randomization
}
\author{
Adam Kapelner
}
