% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_backend.R
\name{compute_multiple_kernel_objective_vals_gpu}
\alias{compute_multiple_kernel_objective_vals_gpu}
\title{Compute multiple kernel objective values using GPU}
\usage{
compute_multiple_kernel_objective_vals_gpu(
  W,
  Kgrams,
  weights,
  initial_objs,
  running_sums,
  max_reds,
  maximum_gain_scaling,
  device = 0
)
}
\arguments{
\item{W}{The design matrix (r x n)}

\item{Kgrams}{A list of Gram matrices}

\item{weights}{A vector of weights for the kernels}

\item{initial_objs}{A vector of initial objective values}

\item{running_sums}{A vector of current kernel sums}

\item{max_reds}{A vector of max reduction log objective values}

\item{maximum_gain_scaling}{Scaling factor}

\item{device}{The device ID}
}
\description{
Compute multiple kernel objective values using GPU
}
