% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_backend.R
\name{full_greedy_search_gpu}
\alias{full_greedy_search_gpu}
\title{Run a full greedy search on GPU (Upload Once, In-Place)}
\usage{
full_greedy_search_gpu(X, Sinv, start_indicT, max_iters = 100, device = 0)
}
\arguments{
\item{X}{The design matrix}

\item{Sinv}{Inverse covariance matrix}

\item{start_indicT}{Starting allocation}

\item{max_iters}{Maximum iterations}

\item{device}{Device ID}
}
\description{
Run a full greedy search on GPU (Upload Once, In-Place)
}
