% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{GreedyExperimentalDesign}
\alias{GreedyExperimentalDesign}
\alias{GreedyExperimentalDesign-package}
\title{Greedy Experimental Design Search}
\description{
A tool to select and run kernel PCA linear models for regression and classification
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}
}
\references{
Kapelner, A
}
\keyword{Design,}
\keyword{Experimental}
\keyword{Greedy}
\keyword{search}

