% Generated by roxygen2 (4.0.1): do not edit by hand
\name{greedySearchCurrentProgress}
\alias{greedySearchCurrentProgress}
\title{Returns the number of vectors found by the greedy design search}
\usage{
greedySearchCurrentProgress(obj)
}
\arguments{
\item{obj}{The \code{greedy_experimental_design} object that is currently running the search}
}
\description{
Returns the number of vectors found by the greedy design search
}
\author{
Adam Kapelner
}

