% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gurobi_numerical_optimization_orthogonalization_curation.R
\name{gurobi_numerical_optimization_orthogonalization_curation}
\alias{gurobi_numerical_optimization_orthogonalization_curation}
\title{Curate More Orthogonal Vectors Via Numerical Optimization}
\usage{
gurobi_numerical_optimization_orthogonalization_curation(
  W,
  R0,
  verbose = FALSE
)
}
\arguments{
\item{W}{A matrix in the set ${-1, 1}^{R x n}$ which have R allocation vectors for an experiment of sample size n.}

\item{R0}{The number of vectors in the pared-down design.}
}
\value{
A list with two elements: (1) \code{avg_abs_rij_by_R} which is a data frame with R - Rmin + 1 rows and 
					columns R and average absolute r_ij and (2) \code{Wsorted} which provides the collection of vectors in
					sorted by best average absolute r_ij in row order from best to worst.
}
\description{
This function takes a set of allocation vectors and pares them down
by eliminating the vectors that can result in the largest reduction in Avg[ |r_ij| ] via
Gurobi's numerical optimization.
}
\details{
Make sure you setup Gurobi properly first. This means applying for a license, downloading, installing, via
\code{install.packages("c:/gurobi902/win64/R/gurobi_9.0-2.zip", repos=NULL)} then make sure you've registered it 
on your computer using the \code{grbgetkey} command with the license file in the default directory.
}
\author{
Adam Kapelner
}
