% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.greedy_experimental_design_search_results}
\alias{plot.greedy_experimental_design_search_results}
\title{Plots a summary of a \code{greedy_experimental_design_search_results} object}
\usage{
\method{plot}{greedy_experimental_design_search}(x, ...)
}
\arguments{
\item{x}{The \code{greedy_experimental_design_search_results} object to be summarized in the plot}

\item{...}{Other parameters to pass to the default plot function}
}
\description{
Plots a summary of a \code{greedy_experimental_design_search_results} object
}
\author{
Adam Kapelner
}

