% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_multiple_kernel_search.R
\name{plot.greedy_multiple_kernel_experimental_design}
\alias{plot.greedy_multiple_kernel_experimental_design}
\title{Plots a summary of a \code{greedy_multiple_kernel_experimental_design} object}
\usage{
\method{plot}{greedy_multiple_kernel_experimental_design}(x, ...)
}
\arguments{
\item{x}{The \code{greedy_multiple_kernel_experimental_design} object to be plotted}

\item{...}{Optional arguments for the \code{plot} function}
}
\description{
Plots a summary of a \code{greedy_multiple_kernel_experimental_design} object
}
\author{
Adam Kapelner
}
