% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_match_search.R
\name{print.binary_experimental_design}
\alias{print.binary_experimental_design}
\title{Prints a summary of a \code{binary_experimental_design} object}
\usage{
\method{print}{binary_experimental_design}(x, ...)
}
\arguments{
\item{x}{The \code{binary_experimental_design} object to be summarized in the console}

\item{...}{Other parameters to pass to the default print function}
}
\description{
Prints a summary of a \code{binary_experimental_design} object
}
\author{
Adam Kapelner
}
