% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_match_search.R
\name{print.binary_match_structure}
\alias{print.binary_match_structure}
\title{Prints a summary of a \code{binary_match_structure} object}
\usage{
\method{print}{binary_match_structure}(x, ...)
}
\arguments{
\item{x}{The \code{binary_match_structure} object to be summarized in the console}

\item{...}{Other parameters to pass to the default print function}
}
\description{
Prints a summary of a \code{binary_match_structure} object
}
\author{
Adam Kapelner
}
