% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerandomization_search.R
\name{print.rerandomization_experimental_design_search}
\alias{print.rerandomization_experimental_design_search}
\title{Prints a summary of a \code{rerandomization_experimental_design_search} object}
\usage{
\method{print}{rerandomization_experimental_design_search}(x, ...)
}
\arguments{
\item{x}{The \code{rerandomization_experimental_design_search} object to be summarized in the console}

\item{...}{Other parameters to pass to the default print function}
}
\description{
Prints a summary of a \code{rerandomization_experimental_design_search} object
}
\author{
Adam Kapelner
}
