% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi_numerical_optimization_search.R
\name{resultsGurobiNumericalOptimizeExperimentalDesign}
\alias{resultsGurobiNumericalOptimizeExperimentalDesign}
\title{Returns the results (thus far) of the Gurobi numerical optimization design search}
\usage{
resultsGurobiNumericalOptimizeExperimentalDesign(obj)
}
\arguments{
\item{obj}{The \code{gurobi_numerical_optimization_experimental_design_search} object that is currently running the search}
}
\description{
Returns the results (thus far) of the Gurobi numerical optimization design search
}
\author{
Adam Kapelner
}
