% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karp_search.R
\name{resultsKarpSearch}
\alias{resultsKarpSearch}
\title{Returns the results (thus far) of the karp design search}
\usage{
resultsKarpSearch(obj)
}
\arguments{
\item{obj}{The \code{karp_experimental_design} object that is currently running the search}
}
\description{
Returns the results (thus far) of the karp design search
}
\examples{
\dontrun{
set.seed(1)
X = matrix(rnorm(10), nrow = 10)
kobj = initKarpExperimentalDesignObject(
  X,
  start = TRUE,
  wait = TRUE,
  balanced = TRUE,
  verbose = FALSE
)
res = resultsKarpSearch(kobj)
res$obj_val
}
}
\author{
Adam Kapelner
}
