% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpsolve_search.R
\name{resultsLPSolveSearch}
\alias{resultsLPSolveSearch}
\title{Returns the results (thus far) of the lPSolve design search}
\usage{
resultsLPSolveSearch(obj)
}
\arguments{
\item{obj}{The \code{lPSolve_experimental_design} object that is currently running the search}
}
\description{
Returns the results (thus far) of the lPSolve design search
}
\author{
Adam Kapelner
}

