% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_search.R
\name{resultsOptimalSearch}
\alias{resultsOptimalSearch}
\title{Returns the results (thus far) of the optimal design search}
\usage{
resultsOptimalSearch(obj, num_vectors = 2, form = "one_zero")
}
\arguments{
\item{obj}{The \code{optimal_experimental_design} object that is currently running the search}

\item{num_vectors}{How many allocation vectors you wish to return. The default is 1 meaning the best vector. If \code{Inf},
it means all vectors.}

\item{form}{Which form should it be in? The default is \code{one_zero} for 1/0's or \code{pos_one_min_one} for +1/-1's.}
}
\description{
Returns the results (thus far) of the optimal design search
}
\examples{
\dontrun{
set.seed(1)
X = matrix(rnorm(12), nrow = 6)
obj = initOptimalExperimentalDesignObject(
  X,
  objective = "abs_sum_diff",
  num_cores = 1,
  start = TRUE,
  wait = TRUE,
  verbose = FALSE
)
res = resultsOptimalSearch(obj, num_vectors = 2, form = "one_zero")
res$opt_obj_val
}
}
\author{
Adam Kapelner
}
