% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{searchTimeElapsed}
\alias{searchTimeElapsed}
\title{Returns the amount of time elapsed}
\usage{
searchTimeElapsed(obj)
}
\arguments{
\item{obj}{The \code{experimental_design} object that is currently running the search}
}
\description{
Returns the amount of time elapsed
}
\examples{
\dontrun{
set.seed(1)
X = matrix(rnorm(20), nrow = 10)
ged = initGreedyExperimentalDesignObject(
  X,
  max_designs = 1,
  num_cores = 1,
  start = TRUE,
  wait = TRUE,
  objective = "abs_sum_diff",
  verbose = FALSE
)
searchTimeElapsed(ged)
}
}
\author{
Adam Kapelner
}
