% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{shuffle_cpp_wrap}
\alias{shuffle_cpp_wrap}
\title{Shuffles a vector rapidly}
\usage{
shuffle_cpp_wrap(w)
}
\arguments{
\item{w}{The vector to be shuffled}
}
\value{
The vector with elements shuffled
}
\description{
Shuffles a vector rapidly
}
\author{
Adam Kapelner
}
