% Generated by roxygen2 (4.0.1): do not edit by hand
\name{startGreedySearch}
\alias{startGreedySearch}
\title{Starts the parallelized greedy design search}
\usage{
startGreedySearch(obj)
}
\arguments{
\item{obj}{The \code{greedy_experimental_design} object that will be running the search}
}
\description{
Starts the parallelized greedy design search
}
\author{
Adam Kapelner
}

