% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{startSearch}
\alias{startSearch}
\title{Starts the parallelized greedy design search.}
\usage{
startSearch(obj)
}
\arguments{
\item{obj}{The \code{experimental_design} object that will be running the search}
}
\description{
Once begun, this function cannot be run again.
}
\examples{
\dontrun{
set.seed(1)
X = matrix(rnorm(20), nrow = 10)
ged = initGreedyExperimentalDesignObject(
  X,
  max_designs = 3,
  num_cores = 1,
  start = FALSE,
  wait = FALSE,
  objective = "abs_sum_diff",
  verbose = FALSE
)
startSearch(ged)
stopSearch(ged)
}
}
\author{
Adam Kapelner
}
