% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stopGreedySearch}
\alias{stopGreedySearch}
\title{Stops the parallelized greedy design search}
\usage{
stopGreedySearch(obj)
}
\arguments{
\item{obj}{The \code{greedy_experimental_design} object that is currently running the search}
}
\description{
Stops the parallelized greedy design search
}
\author{
Adam Kapelner
}

