% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{stopSearch}
\alias{stopSearch}
\title{Stops the parallelized greedy design search.}
\usage{
stopSearch(obj)
}
\arguments{
\item{obj}{The \code{experimental_design} object that is currently running the search}
}
\description{
Once stopped, it cannot be restarted.
}
\examples{
\dontrun{
set.seed(1)
X = matrix(rnorm(20), nrow = 10)
ged = initGreedyExperimentalDesignObject(
  X,
  max_designs = 3,
  num_cores = 1,
  start = TRUE,
  wait = FALSE,
  objective = "abs_sum_diff",
  verbose = FALSE
)
stopSearch(ged)
}
}
\author{
Adam Kapelner
}
