% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_multiple_kernel_search.R
\name{summary.greedy_multiple_kernel_experimental_design}
\alias{summary.greedy_multiple_kernel_experimental_design}
\title{Prints a summary of a \code{greedy_multiple_kernel_experimental_design} object}
\usage{
\method{summary}{greedy_multiple_kernel_experimental_design}(object, ...)
}
\arguments{
\item{object}{The \code{greedy_multiple_kernel_experimental_design} object to be summarized in the console}

\item{...}{Optional arguments for the \code{summary} function}
}
\description{
Prints a summary of a \code{greedy_multiple_kernel_experimental_design} object
}
\author{
Adam Kapelner
}
