% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_search.R
\name{summary.optimal_experimental_design_search}
\alias{summary.optimal_experimental_design_search}
\title{Prints a summary of a \code{optimal_experimental_design_search} object}
\usage{
\method{summary}{optimal_experimental_design_search}(object, ...)
}
\arguments{
\item{object}{The \code{optimal_experimental_design_search} object to be summarized in the console}

\item{...}{Other parameters to pass to the default summary function}
}
\description{
Prints a summary of a \code{optimal_experimental_design_search} object
}
\author{
Adam Kapelner
}
