% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_match_search.R
\name{binaryMatchExperimentalDesignSearch}
\alias{binaryMatchExperimentalDesignSearch}
\title{Begin a Search for Binary Matching Designs}
\usage{
binaryMatchExperimentalDesignSearch(X, compute_dist_matrix = NULL)
}
\arguments{
\item{X}{The design matrix with $n$ rows (one for each subject) and $p$ columns 
(one for each measurement on the subject). This is the design matrix you wish 
to search for a more optimal design.}

\item{compute_dist_matrix}{The function that computes the distance matrix between every two observations in \code{X}, 
its only argument. The default is \code{NULL} signifying euclidean squared distance optimized in C++.}
}
\value{
An object of type \code{binary_experimental_design} which can be further operated upon.
}
\description{
This method creates an object of type binary_experimental_design and will find pairs. You can then
use the function \code{resultsBinaryMatchSearch} to create randomized allocation vectors. For one column
in X, we just sort to find the pairs trivially.
}
\author{
Adam Kapelner
}
