% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_randomization.R
\name{complete_randomization}
\alias{complete_randomization}
\title{Implements complete randomization}
\usage{
complete_randomization(n, r)
}
\arguments{
\item{n}{number of observations}

\item{r}{number of randomized designs you would like}
}
\value{
a matrix where each column is one of the \code{r} designs
}
\description{
Implements complete randomization
}
\author{
Adam Kapelner
}
