% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_randomization_and_blocking.R
\name{complete_randomization_with_forced_balanced}
\alias{complete_randomization_with_forced_balanced}
\title{Implements forced balanced randomization}
\usage{
complete_randomization_with_forced_balanced(n, r, form = "one_zero")
}
\arguments{
\item{n}{number of observations}

\item{r}{number of randomized designs you would like}

\item{form}{Which form should it be in? The default is \code{one_zero} for 1/0's or \code{pos_one_min_one} for +1/-1's.}
}
\value{
a matrix where each column is one of the \code{r} designs
}
\description{
For debugging, you can use \code{set.seed}
to be assured of deterministic output.
}
\author{
Adam Kapelner
}
