% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{compute_distance_matrix_cpp_wrap}
\alias{compute_distance_matrix_cpp_wrap}
\title{Computes a Euclidean-squared distance matrix rapidly}
\usage{
compute_distance_matrix_cpp_wrap(X)
}
\arguments{
\item{X}{A numeric matrix with n rows representing each subject and p columns which are measurements on each subject}
}
\value{
The n x n Euclidean distances squared
}
\description{
Computes a Euclidean-squared distance matrix rapidly
}
\author{
Adam Kapelner
}
