% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{create_all_ys_cpp_wrap}
\alias{create_all_ys_cpp_wrap}
\title{Create all binary Y's convenience function using a randomized design}
\usage{
create_all_ys_cpp_wrap(pCs, pTs, W, two_n, nY)
}
\value{
A matrix of boolean Y's
}
\description{
Create all binary Y's convenience function using a randomized design
}
\author{
Adam Kapelner
}
