% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{gen_pm_designs_cpp_wrap}
\alias{gen_pm_designs_cpp_wrap}
\title{Create PM designs}
\usage{
gen_pm_designs_cpp_wrap(indicies_pairs, n, r)
}
\arguments{
\item{indicies_pairs}{A matrix of n x 2 indicies where each row is a pair of subjects' indicies}

\item{n}{Half the number of subjects i.e. the number of pairs}

\item{r}{The number of assignments to generate}
}
\value{
A matrix of r x 2n PM designs of +1/-1 assignments
}
\description{
Create PM designs
}
\examples{
\dontrun{
indicies_pairs = matrix(c(1, 2, 3, 4), ncol = 2, byrow = TRUE)
gen_pm_designs_cpp_wrap(indicies_pairs, n = 2, r = 3)
}
}
\author{
Adam Kapelner
}
