% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{generate_block_design_cpp_wrap}
\alias{generate_block_design_cpp_wrap}
\title{Generates homogeneous block design allocations rapidly}
\usage{
generate_block_design_cpp_wrap(B, nR, dummy_block)
}
\arguments{
\item{B}{The number of blocks in the design}

\item{nR}{The number of allocation vectors}

\item{dummy_block}{The subvector of allocations in each block that will be permuted}
}
\value{
A matrix with rows being the nR random block allocation of sample size B x length(dummy_block).
}
\description{
Generates homogeneous block design allocations rapidly
}
\author{
Adam Kapelner
}
