% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{generate_stdzied_design_matrix}
\alias{generate_stdzied_design_matrix}
\title{Generates a design matrix with standardized predictors.}
\usage{
generate_stdzied_design_matrix(n = 50, p = 1, covariate_gen = rnorm, ...)
}
\arguments{
\item{n}{Number of rows in the design matrix}

\item{p}{Number of columns in the design matrix}

\item{covariate_gen}{The function to use to draw the covariate realizations (assumed to be iid).
This defaults to \code{rnorm} for $N(0,1)$ draws.}

\item{...}{Optional arguments to be passed to the \code{covariate_dist} function.}
}
\value{
THe design matrix
}
\description{
This function is useful for debugging.
}
\author{
Adam Kapelner
}
