% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate_stdzied_design_matrix}
\alias{generate_stdzied_design_matrix}
\title{Generates a design matrix with standardized predictors. Useful for debugging.}
\usage{
generate_stdzied_design_matrix(n = 50, p = 1,
  covariate_dist = "iid_std_normal")
}
\arguments{
\item{n}{Number of rows in the design matrix}

\item{p}{Number of columns in the design matrix}

\item{covariate_dist}{The distribution of the covariates (defaults to iid_std_normal) and the other option is iid_std_uniform}
}
\value{
THe design matrix
}
\description{
Generates a design matrix with standardized predictors. Useful for debugging.
}
\author{
Adam Kapelner
}

