% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreedyExperimentalDesign.R
\name{GreedyExperimentalDesign}
\alias{GreedyExperimentalDesign}
\title{Greedy Experimental Design Search}
\description{
A tool to find many types of a priori experimental designs
}
\references{
Kapelner, A
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}
}
\keyword{design}
\keyword{optimize}
