% Generated by roxygen2 (4.0.1): do not edit by hand
\name{initGreedyExperimentalDesignObject}
\alias{initGreedyExperimentalDesignObject}
\title{This method creates an object of type greedy_experimental_design and will immediately initiate
a search through $1_{T}$ space.}
\usage{
initGreedyExperimentalDesignObject(X, max_designs = 10000,
  objective = "abs_sum_diff", wait = FALSE, max_iters = NULL,
  semigreedy = FALSE, num_cores = 1)
}
\arguments{
\item{X}{The design matrix with $n$ rows (one for each subject) and $p$ columns
(one for each measurement on the subject). This is the design matrix you wish
to search for a more optimal design.}

\item{max_designs}{The maximum number of designs to be returned. Default is 10,000. Make this large
so you can search however long you wish as the search can be stopped at any time by
using the \code{\link{stopGreedySearch}} method}

\item{objective}{The objective function to use when greedily searching design space. This is a string
"\code{abs_sum_diff}" (default) or "\code{mahal_dist}."}

\item{wait}{Should the \code{R} terminal hang until all \code{max_designs} vectors are found? The
deafult is \code{FALSE}.}

\item{semigreedy}{Should we use a fully greedy approach or the quicker semi-greedy approach? The default is
\code{FALSE} corresponding to the fully greedy approach.}

\item{max_switches}{Should we impose a maximum number of switches? The default is \code{NULL} which a flag
for ``no limit.''}

\item{num_cores}{The number of CPU cores you wish to use during the search. The default is \code{1}.}
}
\value{
An object of type \code{greedy_experimental_design_search} which can be further operated upon
}
\description{
This method creates an object of type greedy_experimental_design and will immediately initiate
a search through $1_{T}$ space.
}
\author{
Adam Kapelner
}

