% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_multiple_kernel_search.R
\name{initGreedyMultipleKernelExperimentalDesignObject}
\alias{initGreedyMultipleKernelExperimentalDesignObject}
\title{Begin A Greedy Multiple Kernel Design Search}
\usage{
initGreedyMultipleKernelExperimentalDesignObject(
  X = NULL,
  kernel_names = NULL,
  Kgrams = NULL,
  kernel_weights = NULL,
  maximum_gain_scaling = TRUE,
  nT = NULL,
  max_designs = 10000,
  kernel_pre_num_designs = 1000,
  wait = FALSE,
  start = TRUE,
  max_iters = Inf,
  semigreedy = FALSE,
  diagnostics = FALSE,
  num_cores = 1,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{X}{The design matrix with $n$ rows (one for each subject) and $p$ columns 
(one for each measurement on the subject). This is the design matrix you wish 
to search for a more optimal design. This parameter must be specified unless you
choose objective type \code{"kernel"} in which case, the \code{Kgrams} parameter must
be specified.}

\item{kernel_names}{A vector of M >= 1 strings indicating the kernels to be used. Valid values are 
"\code{mahalanobis}", "\code{gaussian}", "\code{laplacian}", "\code{exponential}", 
"\code{inv_mult_quad}", "\code{poly_2}", "\code{poly_3}". Or a special value "\code{manual}" to 
indicate the kernels are specified manually using the \code{Kgrams} parameter.}

\item{Kgrams}{A list of M >= 1 elements where each is a \code{n x n} matrix whose
entries are the evaluation of the kernel function between subject i and subject j. Default is \code{NULL}.}

\item{kernel_weights}{A vector of M weights for each kernel which should sum to 1. Default is \code{NULL} which 
specifies equal weighting.}

\item{maximum_gain_scaling}{Should we scale the kernels to have the same maximum gain? Default is \code{TRUE}.}

\item{nT}{The number of treatments to assign. Default is \code{NULL} which is for forced balance allocation
i.e. nT = nC = n / 2 where n is the number of rows in X (or Kgram if X is unspecified).}

\item{max_designs}{The maximum number of designs to be returned. Default is 10,000. Make this large 
so you can search however long you wish as the search can be stopped at any time by
using the \code{\link{stopSearch}} method.}

\item{kernel_pre_num_designs}{The number of initial designs to search through before starting the greedy search for each kernel.
Default is \code{1000}.}

\item{wait}{Should the \code{R} terminal hang until all \code{max_designs} vectors are found? The 
deafult is \code{FALSE}.}

\item{start}{Should we start searching immediately (default is \code{TRUE}).}

\item{max_iters}{Should we impose a maximum number of greedy switches? The default is \code{Inf} which a flag 
for ``no limit.''}

\item{semigreedy}{Should we use a fully greedy approach or the quicker semi-greedy approach? The default is
\code{FALSE} corresponding to the fully greedy approach.}

\item{diagnostics}{Returns diagnostic information about the iterations including (a) the initial starting
vectors, (b) the switches at every iteration and (c) information about the objective function
at every iteration (default is \code{FALSE} to decrease the algorithm's run time).}

\item{num_cores}{The number of CPU cores you wish to use during the search. The default is \code{1}.}

\item{seed}{The set to set for deterministic output. This should only be set if \code{num_cores = 1} otherwise
the output will not be deterministic. Default is \code{NULL} for no seed set.}

\item{verbose}{Should the algorithm emit progress output? Default is \code{TRUE}.}
}
\value{
An object of type \code{greedy_multiple_kernel_experimental_design} which can be further operated upon
}
\description{
This method creates an object of type greedy_multiple_kernel_experimental_design and will immediately initiate
a search through allocation space for forced balance designs. For debugging, you can use set the \code{seed}
parameter and \code{num_cores = 1} to be assured of deterministic output.
}
\author{
Adam Kapelner
}
