% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_obj_val_by_iter}
\alias{plot_obj_val_by_iter}
\title{Plots the objective value by iteration}
\usage{
plot_obj_val_by_iter(res, runs = NULL)
}
\arguments{
\item{res}{Results from a greedy search object}

\item{runs}{A vector of run indices you would like to see plotted (default is to plot the first up to 9)}
}
\description{
Plots the objective value by iteration
}
\author{
Adam Kapelner
}

