% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_multiple_kernel_search.R
\name{print.greedy_multiple_kernel_experimental_design}
\alias{print.greedy_multiple_kernel_experimental_design}
\title{Prints a summary of a \code{greedy_multiple_kernel_experimental_design} object}
\usage{
\method{print}{greedy_multiple_kernel_experimental_design}(x, ...)
}
\arguments{
\item{x}{The \code{greedy_multiple_kernel_experimental_design} object to be summarized in the console}

\item{...}{Other parameters to pass to the default print function}
}
\description{
Prints a summary of a \code{greedy_multiple_kernel_experimental_design} object
}
\author{
Adam Kapelner
}
