% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_match_search.R
\name{resultsBinaryMatchSearch}
\alias{resultsBinaryMatchSearch}
\title{Binary Matched vectors}
\usage{
resultsBinaryMatchSearch(
  obj,
  num_vectors = 1000,
  objective = NULL,
  form = "zero_one",
  prop_flips = 1
)
}
\arguments{
\item{obj}{The \code{binary_experimental_design} object where the pairs are computed.}

\item{num_vectors}{How many random allocation vectors you wish to return. The default is 1000.}

\item{objective}{Should we compute all the objective values for each allocation? Default is \code{NULL} for "no".
If non-null, it needs to either be "mahal_dist" or "abs_sum_diff".}

\item{form}{Which form should it be in? The default is \code{one_zero} for 1/0's or \code{pos_one_min_one} for +1/-1's.}

\item{prop_flips}{Proportion of flips. Default is all. Lower for more correlated assignments (useful for research only).}
}
\description{
Returns unique allocation vectors that are binary matched. For debugging, you can use \code{set.seed}
to be assured of deterministic output.
}
\author{
Adam Kapelner
}
