% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_match_search.R
\name{resultsBinaryMatchSearch}
\alias{resultsBinaryMatchSearch}
\title{)
Returns allocation vectors that are binary matched}
\usage{
resultsBinaryMatchSearch(obj, num_vectors = 1000)
}
\arguments{
\item{obj}{The \code{binary_experimental_design} object where the pairs are computed.}

\item{num_vectors}{How many random allocation vectors you wish to return. The default is 1000.}
}
\description{
)
Returns allocation vectors that are binary matched
}
\author{
Adam Kapelner
}
