% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resultsGreedySearch}
\alias{resultsGreedySearch}
\title{Returns the results (thus far) of the greedy design search}
\usage{
resultsGreedySearch(obj, max_vectors = NULL)
}
\arguments{
\item{obj}{The \code{greedy_experimental_design} object that is currently running the search}

\item{max_vectors}{The number of design vectors you wish to return. \code{NULL} returns all of them
                        (default). This is not recommended as returning over 1,000 vectors is time-intensive.}
}
\description{
Returns the results (thus far) of the greedy design search
}
\author{
Adam Kapelner
}

