% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karp_search.R
\name{resultsKarpSearch}
\alias{resultsKarpSearch}
\title{Returns the results (thus far) of the karp design search}
\usage{
resultsKarpSearch(obj)
}
\arguments{
\item{obj}{The \code{karp_experimental_design} object that is currently running the search}
}
\description{
Returns the results (thus far) of the karp design search
}
\author{
Adam Kapelner
}
