% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_multiple_kernel_search.R
\name{resultsMultipleKernelGreedySearch}
\alias{resultsMultipleKernelGreedySearch}
\title{Returns the results of a greedy multiple kernel search}
\usage{
resultsMultipleKernelGreedySearch(obj, max_vectors = NULL, form = "one_zero")
}
\arguments{
\item{obj}{The \code{greedy_multiple_kernel_experimental_design} object where the search was run.}

\item{max_vectors}{How many random allocation vectors you wish to return. The default is \code{NULL} indicating you want all of them.}

\item{form}{Which form should it be in? The default is \code{one_zero} for 1/0's or \code{pos_one_min_one} for +1/-1's.}
}
\description{
Returns the results of a greedy multiple kernel search
}
\author{
Adam Kapelner
}
