% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resultsOptimalSearch}
\alias{resultsOptimalSearch}
\title{Returns the results (thus far) of the optimal design search}
\usage{
resultsOptimalSearch(obj)
}
\arguments{
\item{obj}{The \code{optimal_experimental_design} object that is currently running the search}
}
\description{
Returns the results (thus far) of the optimal design search
}
\author{
Adam Kapelner
}

