% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerandomization_search.R
\name{resultsRerandomizationSearch}
\alias{resultsRerandomizationSearch}
\title{Returns the results (thus far) of the rerandomization design search}
\usage{
resultsRerandomizationSearch(
  obj,
  include_assignments = FALSE,
  form = "one_zero"
)
}
\arguments{
\item{obj}{The \code{rerandomization_experimental_design} object that is currently running the search}

\item{include_assignments}{Do we include the assignments (takes time) and default is \code{FALSE}.}

\item{form}{Which form should the assignments be in? The default is \code{one_zero} for 1/0's or \code{pos_one_min_one} for +1/-1's.}
}
\description{
Returns the results (thus far) of the rerandomization design search
}
\author{
Adam Kapelner
}
