% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rerandomization_search.R
\name{resultsRerandomizationSearch}
\alias{resultsRerandomizationSearch}
\title{Returns the results (thus far) of the rerandomization design search}
\usage{
resultsRerandomizationSearch(obj)
}
\arguments{
\item{obj}{The \code{rerandomization_experimental_design} object that is currently running the search}
}
\description{
Returns the results (thus far) of the rerandomization design search
}
\author{
Adam Kapelner
}

