% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{searchTimeElapsed}
\alias{searchTimeElapsed}
\title{Returns the number of vectors found by the greedy design search}
\usage{
searchTimeElapsed(obj)
}
\arguments{
\item{obj}{The \code{experimental_design} object that is currently running the search}
}
\description{
Returns the number of vectors found by the greedy design search
}
\author{
Adam Kapelner
}
