% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{standardize_data_matrix}
\alias{standardize_data_matrix}
\title{Standardizes the columns of a data matrix.}
\usage{
standardize_data_matrix(X)
}
\arguments{
\item{X}{The n x p design matrix}
}
\value{
The n x p design matrix with columns standardized
}
\description{
Standardizes the columns of a data matrix.
}
\author{
Adam Kapelner
}
