% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{startSearch}
\alias{startSearch}
\title{Starts the parallelized greedy design search.}
\usage{
startSearch(obj)
}
\arguments{
\item{obj}{The \code{experimental_design} object that will be running the search}
}
\description{
Once begun, this function cannot be run again.
}
\author{
Adam Kapelner
}
