% Generated by roxygen2 (4.0.1): do not edit by hand
\name{startSearch}
\alias{startSearch}
\title{Starts the parallelized greedy design search. Once begun, this function cannot be run again.}
\usage{
startSearch(obj)
}
\arguments{
\item{obj}{The \code{greedy_experimental_design} object that will be running the search}
}
\description{
Starts the parallelized greedy design search. Once begun, this function cannot be run again.
}
\author{
Adam Kapelner
}

