% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{stopSearch}
\alias{stopSearch}
\title{Stops the parallelized greedy design search.}
\usage{
stopSearch(obj)
}
\arguments{
\item{obj}{The \code{experimental_design} object that is currently running the search}
}
\description{
Once stopped, it cannot be restarted.
}
\author{
Adam Kapelner
}
