% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stopSearch}
\alias{stopSearch}
\title{Stops the parallelized greedy design search. Once stopped, it cannot be restarted.}
\usage{
stopSearch(obj)
}
\arguments{
\item{obj}{The \code{experimental_design} object that is currently running the search}
}
\description{
Stops the parallelized greedy design search. Once stopped, it cannot be restarted.
}
\author{
Adam Kapelner
}

